+++ examples/multi_material_deferred.py
@@ -29,7 +29,8 @@ gbuff = t3.FrameBuffer(camera, buffers=dict(
     position=[3, float],
     texcoord=[2, float],
     normal=[3, float],
-    tangent=[3, float]))
+    tangent=[3, float],
+))
 imgbuf = t3.DeferredShading(gbuff, mtllib)
 scene.add_buffer(imgbuf)
 light = t3.Light([0, -0.5, -1], 0.9)
@@ -37,7 +38,7 @@ scene.add_light(light)
 ambient = t3.AmbientLight(0.1)
 scene.add_light(ambient)
 
-gui = ti.GUI('Model', camera.res)
+gui = ti.GUI('Deferred materials', camera.res)
 while gui.running:
     gui.get_event(None)
     gui.running = not gui.is_pressed(ti.GUI.ESCAPE)