+++ taichi_three/common.py
@@ -97,12 +97,3 @@ class dummy_expression:
             return self
         wrapped.__name__ = key
         return wrapped
-
-_old_begin_frontend_if = ti.begin_frontend_if
-
-def _begin_frontend_if(cond):
-    if isinstance(cond, dummy_expression):
-        cond = 0
-    return _old_begin_frontend_if(cond)
-
-ti.begin_frontend_if = _begin_frontend_if